fix: successRate를 Double로 받도록 수정

+++ android/app/src/main/java/com/created/team201/presentation/studyDetail/StudyDetailViewModel.kt
@@ -111,7 +111,7 @@ class StudyDetailViewModel(
             isApplicant = isApplicant,
             profileImageUrl = this.profileImage,
             name = this.nickname,
-            successRate = this.successRate,
+            successRate = this.successRate.toInt(),
             tier = this.tier,
         )
